requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
cycler==0.11.0
dnspython==2.2.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.1
greenlet==1.1.2
is-disposable-email==1.0.0
is-safe-url==1.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
msgpack==1.0.3
mysql-connector-python==8.0.29
numpy==1.22.1
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pycparser==2.21
pynvim==0.4.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
wtforms-validators==1.0.0
